+++ tests/testthat/test_data_valid.R
@@ -9,12 +9,12 @@ test_that(desc = "Test the dates", {
 
 test_that(desc = "Test the structure", {
 
-  expect_equal(base::ncol(italy_total) == 12, TRUE)
-  expect_equal(base::ncol(italy_region) == 17, TRUE)
+  expect_equal(base::ncol(italy_total) == 14, TRUE)
+  expect_equal(base::ncol(italy_region) == 19, TRUE)
   expect_equal(base::ncol(italy_province) == 11, TRUE)
 
-  expect_equal(base::nrow(italy_total) >= 58, TRUE)
-  expect_equal(base::nrow(italy_region) >= 1200, TRUE)
-  expect_equal(base::nrow(italy_province) >= 3072, TRUE)
+  expect_equal(base::nrow(italy_total) >= 140, TRUE)
+  expect_equal(base::nrow(italy_region) >= 3000, TRUE)
+  expect_equal(base::nrow(italy_province) >= 18000, TRUE)
 
 })